Fix Github login
Pro tip: For more information about the oAuth flow and strategy see the oAuth API documentation.

When we set up the authentication client in the browser it can also already handle oAuth logins. To log in with GitHub, visit localhost:3030/oauth/github. It will redirect to GitHub and ask to authorize our application. If everything went well, you will see a user with your GitHub email address being logged in the console.

Note: The authentication client will not use the token from the oAuth login if there is already another token logged in. See the oAuth API for how to link to an existing account
Co-Authored-By: Edward Almanzar <edward@menteware.com>