Buidling frontend - Plain JavaScript real-time chat frontend with login and signup
That’s it. We now have a plain JavaScript real-time chat frontend with login and signup. This example demonstrates many of the basic principles of how you interact with a Feathers API. You can log in with your GitHub account by following the "Login with GitHub" button, with the email (hello@feathersjs.com) and password (supersecret) from the user we registered in the authentication chapter or sign up and log in with a different email address.

Co-Authored-By: Edward Almanzar <edward@menteware.com>